+++ app/src/client/components/emulator/index.js
@@ -1,2 +1,2 @@
 export {default as Emulator} from './Emulator';
-export {default as EmulatorToolbar} from './EmulatorToolbar';
+export {EmulatorToolbar} from './toolbar';